refactor(US-006): Extract endpoint response models to separate files
- Created Endpoints/Models directory
- Extracted GetMyRolesResponse to separate file
- Extracted RoleInfo to separate file
- Extracted RequestGameMasterRoleResponse to separate file
- Updated RoleEndpoints.cs to use Models namespace
- Improved XML documentation with parameters and return types
- Follows .NET convention: one public type per file
